Use credentials file passed through config (#69)
* Use credentials file passed through config

Add a new supported key to the config which allows the path to a
credentials file to be passed through to the ObjectStore plugin.
If that key is set in the config, use that file as the credentials
file for the AWS client.

This can be tested by adding a new key to the `cloud-credentials` secret
and adding the key value pair `credentialsFile=/credentials/<key_name>`.

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>

* Add function to create session options

We have some repeated logic for creating session option. This introduces
a new function to handle the creation of those session options.

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>

* Add documentation for per-BSL credentials

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>

* Address code review and add changelog

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>