Load AWS credentials from Kubernetes secret
When Velero is using AWS IRSA for authentication, it is
hard to use another provider in the chain, because the IRSA Role is
always picked before a credentials file or static credentials in the
environment.
To mitigate this, add an extra configuration option `KubernetesSecret`,
when this option is set, static credentials will be loaded from a k8s
secret with that name. When the option is not set, the default
CredentialsProviderChain is used.

Signed-off-by: Rob Kenis <robke@softwareag.com>